Reduce container size by fetching github zip archives
Instead of running `git clone` for our 4 code repositories, now we download
the zip archives instead. This avoids the extra space taken up by the .git
directories and also avoids downloading unnecessary history (as an
alternative to `git clone --depth 1`).

We also use ADD statements on the github API for commits as a cache buster
for updates to any of the repositories.